Generic servo tuning ; fix bug in robot.moveRail

+++ eurobot2021/MainRobotCode/embedded/include/uCListener.h
@@ -17,7 +17,7 @@
     static inline std::ostream& operator<< (std::ostream& out, ExcavationSquareColor const& c)
     {
         std::string const names[] = {"None", "Purple", "Yellow", "Red"};
-        return out << names[c];
+        return out << names[static_cast<int>(c)];
     }
